Initial Commit
This version of the XML Demo has a pretty bare fake computational
engine that doesn’t look at inputs.xml and generates random outputs.
The GUI in JavaAppDemo has unused fields for inputs, triggers the
computational engine when “Trigger is pressed”, and updates the outputs
from “outputs.xml” when “View Outputs is pressed”. The projects were
made with Netbeans and includes a nice way to visually design the GUI
by drag and drop. See the tutorial I will post for that.